refactor(scheduler): extract triple and quality measurement handlers (issue #224 task 5)
